+++ Sourse/Application/Log.cpp
