+++ Sourse/Types/Parser.cpp
@@ -1,5 +1,5 @@
-#include "public\FileReader.h"
-#include "Mem.h"
+#include "public/Parser.h"
+#include "public/Mem.h"
 
 #include <cassert>
 #include <fstream>
